commit update dashboard estrutura
