--- client/src/app/api/users/route.ts
@@ -1,31 +1,0 @@
-import { NextResponse } from "next/server";
-import { PrismaClient } from "../../../generated/prisma";
-
-const prisma = new PrismaClient();
-
-export async function GET() {
-  try {
-    const users = await prisma.user.findMany();
-    console.log("Usuários encontrados:", users);
-    return NextResponse.json(users);
-  } catch (error) {
-    return NextResponse.json({ error: String(error) }, { status: 500 });
-  }
-}
-
-export async function POST(req: Request) {
-  try {
-    const body = await req.json();
-
-    const user = await prisma.user.create({
-      data: {
-        name: body.name,
-        email: body.email,
-      },
-    });
-
-    return NextResponse.json(user);
-  } catch (error) {
-    return NextResponse.json({ error: String(error) }, { status: 500 });
-  }
-}
